implemented filter options for --search an package filter option
